scale img

+++ app/src/main/java/union/uc/com/rxjava_example/ui/Utils.java
@@ -8,8 +8,8 @@ import android.util.TypedValue;
  * Created by wangli on 4/17/16.
  */
 public class Utils {
-  public static float dipToPixels(Context context, float dipValue) {
+  public static int dipToPixels(Context context, int dipValue) {
     DisplayMetrics metrics = context.getResources().getDisplayMetrics();
-    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dipValue, metrics);
+    return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dipValue, metrics);
   }
 }
